requirements.txt
discord.py==2.4.1
APScheduler==3.10.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
openai==1.29.0
